chore: MVVM Implemented
