+++ moon-store-mac-os/moon-store-mac-os/Networking/MSEndpoint.swift
