+++ moon-store-mac-os/moon-store-mac-os/Networking/Models/RoleModel.swift
