+++ moon-store-mac-os/moon-store-mac-os/Views/Main/Components/Sidebar.swift
@@ -10,7 +10,7 @@ import SwiftUI
 struct Sidebar: View {
     @EnvironmentObject private var router: AppRouter
     @Binding private var screenSelection: Screen
-    private let authManager: AuthenticationManager = .init()
+    private let loginViewModel: LoginViewModel = .init()
     
     init(screenSelection: Binding<Screen>) {
         _screenSelection = screenSelection
@@ -56,7 +56,7 @@ struct Sidebar: View {
     
     private var logoutButton: some View {
         Button {
-            authManager.logout()
+            loginViewModel.logout()
             router.pop()
         } label: {
             Label("Cerrar Sesión", systemImage: "person.badge.minus")
